pyproject.toml
[tool.poetry]
name = "glomeruli-detection"
version = "0.1.0"
description = "A tool for detecting glomeruli in medical images"
authors = ["Manuel Escobar Ferrer <manuelescobar.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
matplotlib = "^3.9.1"
numpy = "<2"
torch = {version = "^2.3.1", source = "pytorch"}
torchvision = {version = "^0.18.1", source = "pytorch"}
torchsummary = "^1.5.1"
openslide-python = "^1.3.1"
shapely = "^2.0.4"
h5py = "^3.11.0"
lmdb = "^1.5.1"
opencv-python = "^4.10.0.84"
tqdm = "^4.66.4"
scikit-image = "^0.24.0"
pandas = "^2.2.2"
scikit-learn = "^1.5.1"
umap-learn = "^0.5.6"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"